Simplify Section and move into profile.py
